H3: transferencia

+++ src/main/java/com/keepu/webAPI/dto/request/CreateTransactionRequest.java
@@ -6,6 +6,8 @@ import com.keepu.webAPI.model.Wallet;
 import com.keepu.webAPI.model.enums.TransactionType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CreateTransactionRequest(
@@ -14,7 +16,7 @@ public record CreateTransactionRequest(
         Wallet wallet,
 
         @NotNull(message = "El monto no puede ser nulo")
-        Double amount,
+        BigDecimal amount,
 
         @NotBlank(message = "La descripción no puede estar vacía")
         String description,
